requirements.txt
adjustText
alembic
altair
anthropic>=0.3.1
datasets
fastapi==0.100.0
fire
hydra-core
matplotlib
openai==0.28.0
pandas
polars
pre-commit
psycopg2-binary==2.9.6
pydantic<2.0.0 # above 2 breaks anthropic
scikit-learn
scipy
seaborn
sqlalchemy==2.0.18
tabulate
tenacity
termcolor
tiktoken
tqdm
trueskill
typer[all]
uvicorn[standard]==0.22.0
wandb
